+++ dt-connector/src/extractor/redis/rdb/reader/list_pack.rs
@@ -85,31 +85,31 @@ impl RdbReader<'_> {
             // 13bit int
             let second_byte = reader.read_u8()?;
             uval = (u64::from(first_byte & 0x1f) << 8) | u64::from(second_byte); // 5bit + 8bit, 0x1f is 00011111
-            negstart = (1 as u64) << 12;
+            negstart = (1_u64) << 12;
             negmax = 8191; // uint13_max
             let _ = reader.read_bytes(Self::lp_encode_backlen(2));
         } else if (first_byte & LP_ENCODING_16BIT_INT_MASK) == LP_ENCODING_16BIT_INT {
             // 16bit int
             uval = reader.read_u16::<LittleEndian>()? as u64;
-            negstart = (1 as u64) << 15;
+            negstart = (1_u64) << 15;
             negmax = u16::MAX as u64;
             let _ = reader.read_bytes(Self::lp_encode_backlen(2)); // encode: 1byte, int: 2
         } else if (first_byte & LP_ENCODING_24BIT_INT_MASK) == LP_ENCODING_24BIT_INT {
             // 24bit int
             uval = reader.read_u24::<LittleEndian>()? as u64;
-            negstart = (1 as u64) << 23;
+            negstart = (1_u64) << 23;
             negmax = (u32::MAX >> 8) as u64; // uint24_max
             let _ = reader.read_bytes(Self::lp_encode_backlen(1 + 3)); // encode: 1byte, int: 3byte
         } else if (first_byte & LP_ENCODING_32BIT_INT_MASK) == LP_ENCODING_32BIT_INT {
             // 32bit int
             uval = reader.read_u32::<LittleEndian>()? as u64;
-            negstart = (1 as u64) << 31;
+            negstart = (1_u64) << 31;
             negmax = u32::MAX as u64; // uint32_max
             let _ = reader.read_bytes(Self::lp_encode_backlen(1 + 4)); // encode: 1byte, int: 4byte
         } else if (first_byte & LP_ENCODING_64BIT_INT_MASK) == LP_ENCODING_64BIT_INT {
             // 64bit int
             uval = reader.read_u64::<LittleEndian>()?;
-            negstart = (1 as u64) << 63;
+            negstart = (1_u64) << 63;
             negmax = u64::MAX; // uint64_max
             let _ = reader.read_bytes(Self::lp_encode_backlen(1 + 8)); // encode: 1byte, int: 8byte
         } else if (first_byte & LP_ENCODING_12BIT_STR_MASK) == LP_ENCODING_12BIT_STR {
